clean up configs
