+++ scripts/deploy/osmosis/mainIndex.ts
@@ -1,5 +1,5 @@
 import { taskRunner } from '../base'
-import { osmosisMainnet } from './config.js'
+import { osmosisMainnet } from './testnetConfig'
 
 void (async function () {
   await taskRunner(osmosisMainnet)
