+++ scripts/deploy/osmosis/multisig.ts
@@ -1,5 +1,5 @@
 import { taskRunner } from '../base'
-import { osmosisTestMultisig } from './config.js'
+import { osmosisTestMultisig } from './testnetConfig'
 
 void (async function () {
   await taskRunner(osmosisTestMultisig)
