+++ scripts/deploy/osmosis/testIndex.ts
@@ -1,5 +1,5 @@
 import { taskRunner } from '../base'
-import { osmosisTestnetConfig } from './config.js'
+import { osmosisTestnetConfig } from './testnetConfig'
 
 void (async function () {
   await taskRunner(osmosisTestnetConfig)
